This will completely fuck over Travis. More info in description.
This is a necessary update, we will find a new solution to check the syntax of all modules of the bot once ES6 is added to the entire codebase. This update adds helpful logging to understand what the bot's connection process is as well as provides new access to the development database account on our development environments.